Rename old JSONGroup
